refactor(models): move model code to models package Signed-off-by: Cameron Smith <cameron.ray.smith@gmail.com>
