+++ src/pyrovelocity/models/__init__.py
@@ -1,0 +1,4 @@
+from pyrovelocity.models._velocity import PyroVelocity
+
+
+__all__ = [PyroVelocity]
