+++ src/pyrovelocity/models/_transcription_dynamics.py
