+++ src/pyrovelocity/tests/test_velocity.py
@@ -2,6 +2,6 @@
 
 
 def test_load__velocity():
-    from pyrovelocity import _velocity
+    from pyrovelocity.models import _velocity
 
     print(_velocity.__file__)
