+++ src/pyrovelocity/tests/test_velocity_module.py
@@ -2,6 +2,6 @@
 
 
 def test_load__velocity_module():
-    from pyrovelocity import _velocity_module
+    from pyrovelocity.models import _velocity_module
 
     print(_velocity_module.__file__)
