+++ src/pyrovelocity/train.py
@@ -13,10 +13,10 @@ from beartype import beartype
 from mlflow import MlflowClient
 from numpy import ndarray
 
-from pyrovelocity._velocity import PyroVelocity
 from pyrovelocity.data import load_anndata_from_path
 from pyrovelocity.io.compressedpickle import CompressedPickle
 from pyrovelocity.logging import configure_logging
+from pyrovelocity.models import PyroVelocity
 from pyrovelocity.utils import print_anndata, print_attributes
 
 logger = configure_logging(__name__)
